+++ test/e2e/tests/confirmations/transactions/speed-up-and-cancel-confirmations.spec.ts
@@ -5,7 +5,7 @@ import { login } from '../../../page-objects/flows/login.flow';
 import { DEFAULT_FIXTURE_ACCOUNT, WINDOW_TITLES } from '../../../constants';
 import { withFixtures } from '../../../helpers';
 import FixtureBuilderV2 from '../../../fixtures/fixture-builder-v2';
-import { createDappTransaction } from '../../../page-objects/flows/transaction';
+import { createDappTransaction } from '../../../page-objects/flows/transaction.flow';
 import Confirmation from '../../../page-objects/pages/confirmations/confirmation';
 import SpeedUpAndCancelModal from '../../../page-objects/pages/confirmations/speed-up-and-cancel-modal';
 import ActivityListPage from '../../../page-objects/pages/home/activity-list';
