+++ test/e2e/tests/transaction/edit-gas-fee.spec.ts
@@ -3,7 +3,7 @@ import { login } from '../../page-objects/flows/login.flow';
 import {
   createInternalTransaction,
   createDappTransaction,
-} from '../../page-objects/flows/transaction';
+} from '../../page-objects/flows/transaction.flow';
 import { WINDOW_TITLES } from '../../constants';
 import { withFixtures } from '../../helpers';
 import FixtureBuilderV2 from '../../fixtures/fixture-builder-v2';
