+++ test/e2e/tests/transaction/gas-estimates.spec.ts
@@ -1,6 +1,6 @@
 import { MockttpServer } from 'mockttp';
 import { login } from '../../page-objects/flows/login.flow';
-import { createInternalTransaction } from '../../page-objects/flows/transaction';
+import { createInternalTransaction } from '../../page-objects/flows/transaction.flow';
 import { withFixtures } from '../../helpers';
 import FixtureBuilderV2 from '../../fixtures/fixture-builder-v2';
 import { CHAIN_IDS } from '../../../../shared/constants/network';
